+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file is a history of the changes made to @idearium/cli.
 
+## Unreleased
+
+-   New commands.
+
+### Improvements
+
+    Added the `c kc manifests` command.
+
 ## v3.0.2 (5 August 2019)
 
 -   Improved commands.